Uploading Images to Supabase and Editing a Cabin

+++ src/services/apiCabins.js
@@ -1,3 +1,4 @@
+import { id } from "date-fns/locale";
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
@@ -10,24 +11,40 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     "",
   );
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1) Create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([
+  // 1) Create/Edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id)
+    query = query.insert([
       {
         ...newCabin,
         image: imagePath,
       },
-    ])
-    .select();
+    ]);
+
+  // B) EDIT
+  if (id)
+    query = query
+      .update({
+        ...newCabin,
+        image: imagePath,
+      })
+      .eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -35,18 +52,20 @@ export async function createCabin(newCabin) {
   }
 
   // 2) Upload image
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  // 3) If image upload fails, delete cabin
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
 
-    console.error(storageError);
-    throw new Error(
-      "cabins Image could not be uploaded and the cabin was not created",
-    );
+      console.error(storageError);
+
+      throw new Error(
+        "Cabin image could not be uploaded and the cabin was not created",
+      );
+    }
   }
 
   return data;
